code refactor
remove all lock, use channel instead

+++ split.go
@@ -1,7 +1,7 @@
 package btree
 
 //split leaf
-func (l *Leaf) split(tree *Btree) (key []byte, left, right int32) {
+func (l *Leaf) split(tree *Btree) (key []byte, left, right int64) {
 	nleaf := tree.newLeaf()
 	mid := tree.GetLeafMax() / 2
 	nleaf.Values = make([][]byte, len(l.Values[mid:]))
@@ -17,14 +17,14 @@ func (l *Leaf) split(tree *Btree) (key []byte, left, right int32) {
 }
 
 //split node
-func (n *Node) split(tree *Btree) (key []byte, left, right int32) {
+func (n *Node) split(tree *Btree) (key []byte, left, right int64) {
 	nnode := tree.newNode()
 	mid := tree.GetNodeMax() / 2
 	key = n.Keys[mid]
 	nnode.Keys = make([][]byte, len(n.Keys[mid+1:]))
 	copy(nnode.Keys, n.Keys[mid+1:])
 	n.Keys = n.Keys[:mid]
-	nnode.Childrens = make([]int32, len(n.Childrens[mid+1:]))
+	nnode.Childrens = make([]int64, len(n.Childrens[mid+1:]))
 	copy(nnode.Childrens, n.Childrens[mid+1:])
 	n.Childrens = n.Childrens[:mid+1]
 	left = n.GetId()
